Refactoring components
Added components folder

Move app.js to folder
Added DrumBox.js and Knob.js with component
