+++ src/components/Knob.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+
+const Knob = () => {
+  return (
+    <div class="large-knob">
+      <div class="large-knob-shape shapeGR">
+        <div class="large-knob-beveler">
+          <div class="large-knob-marker"></div>
+        </div>
+      </div>
+      <div class="svg-wrapp"></div>
+    </div>
+  )
+}
+
+export default Knob
